lksctp-tools: update and remove from blacklist
Switching lksctp-tools to build from the upstream git repository rather
than a source tarball.  At the same time we'll un-blacklist it since it
appears the gold-as-ld issue has been resolved underneath us, probably
related to a Debian bug of the same form that was also resolved with a
binutils update.

Signed-off-by: Joe MacDonald <joe_macdonald@mentor.com>
Signed-off-by: Armin Kuster <akuster808@gmail.com>